Slight tweak to folder structure to improve flexibility.
In the initial release of Wee_ users had to use `md` files, otherwise
content wouldn't render. MarkDown has quite a few different extensions
and I wanted to actually include support for them.

In config.php you can now choose your extension for `post` and `index`
files. The default is set to `md`, and the files in the initial
download match this. But it is now possible to change this default to
match up with what you are accustomed to.

These changes will also affect the error pages listed in the `/errors/`
folder, but an extra contingency has been built in so that if you don't
want to bother changing these files you won't have to, it will always
fall back to `md` files if you don't edit these yourself.