Date Formatting for PHP < 5.3
I've made a tweak to how dates are formatted for use in the site, to
include older versions of PHP. The old version was using a function
introduced in PHP 5.3, which rendered an important element of Wee_
unusable to many people that might want to use it.

This has now been changed to include those users.