Better comment saving
Wee_ can now be set to not save questionable comments. This has been
implemented to prevent servers from being filled with spam files, a
potential issue with flat-file systems.

By default the config.php is now set to save these comments, but can be
changed so that they are *not* saved. To do so change:

define('saveSpam',true);

to the following:

define('saveSpam',false);

This change has been implemented in the two themes that I currently
provide that have commenting functionality, the starkers theme as well
as my portfolio theme.