Post Tags
Adding tags in the post meta of a post will now add tags to the post if
you have a [[tags]] field in your content.php file.

Additionally, you can now look for posts with a specific tag, or link
to a specific tag on the site by going to /tag/%tag you want%.

This feature is courtesy of Chris Grant, @duckbox on twitter.